requirements.txt
langchain
langchain-ollama
langchain-community
langchain-core
openai-whisper
streamlit
python-dotenv